refactor: update package structure for custom GSON adapters to `controller.adapters`
