+++ core/src/main/java/io/github/com/ranie_borges/thejungle/controller/adapters/AmbientAdapter.java
@@ -1,4 +1,4 @@
-package io.github.com.ranie_borges.thejungle.controller.systems;
+package io.github.com.ranie_borges.thejungle.controller.adapters;
 
 import com.google.gson.*;
 import io.github.com.ranie_borges.thejungle.model.world.Ambient;
